Fix a nasty bug with ArrayFire being doubly linked, enforce consistency across fl/w2l
Summary:
(long, but worth the read if you want to understand what was happening)

Before, in both flashlight and wav2letter++, ArrayFire was linked in multiple ways. One way was the newer `imported` target (`ArrayFire::afcuda`, `ArrayFire::afcpu`, etc), which automatically set library and include paths on a target and was backend-specific. Before this feature was available/tested, a legacy CMake variable was used, `${ArrayFire_LIBRARIES}`, which still exists in ArrayFire for backwards compatibility purposes with older versions of CMake that don't support imported targets.

Since the flashlight install step created a CMake config that required all targets to be independently exported, it was easier to keep `${ArrayFire_LIBRARIES}` rather than explicitly specifying the correct ArrayFire target to export, because the path would be tacked onto any existing library paths for flashlight.

The problem was that `${ArrayFire_LIBRARIES}` included concatenated paths to *all* ArrayFire backend libs. The effect of this was that if the user had multiple versions of ArrayFire installed, multiple ArrayFire libraries, each with the same symbols, would be linked into flashlight or wav2letter++. This is especially sinister because ArrayFire tries to install as many backends as possible; if a user has `libcuda` but doesn't have a CUDA driver installed (or even a GPU connected to their machine!), libafcuda can still be built, and would be linked to fl/w2l vis-a-vis `${ArrayFire_LIBRARIES}`. In this case, random ArrayFire functions from flashlight/wav2letter++ will transitively call into libcuda, which has undefined symbols.

The linker never complained about this (gold, ld, or lld), across operating systems, and silently assigned some functions symbols in `afcpu`, and others symbols in `afcuda`. This had many disparate and seemingly unrelated effects, including:
- Using a `DevicePtr` with the flashlight CPU backend would always segfault, but not until the pointer was read/written to. If the ArrayFire CUDA backend was built/linked, getting a CPU device pointer would give a pointer to memory on the GPU. Whenever any IO to the pointer's virtual device memory occurred, things would segfault, because ArrayFire didn't know how to resolve mutations to GPU memory made through CPU backend symbols.
- Users who had built multiple ArrayFire backends might see random build failures, even if they had the flashlight/w2l CPU backend selected. ArrayFire would build the CUDA backend via libcuda, but might not be able to link to other CUDA symbols (e.g. cuSolver), so CUDA linker errors would occur (for example, https://github.com/facebookresearch/wav2letter/issues/108)
- Random runtime errors when using wav2letter++ with different criterion backends, due to arrays randomly being allocated/read across various devices (for example, https://github.com/facebookresearch/wav2letter/issues/111)

Implements the following simple fixes:
- Removes all instances of `${ArrayFire_LIBRARIES}` across all build configs, and exclusively uses the ArrayFire imported target
- Only links an imported ArrayFire target once across all of flashlight and w2l.
- wav2letter++ now uses the imported ArrayFire target used to build flashlight which is exported by flashlight and transitively included.

Reviewed By: vineelpratap

Differential Revision: D13567348

fbshipit-source-id: 5c4533035b8c3a033b706cb37dc38dcd6470fcfc