Build support for CUDA/NVCC, CUB, and criterion kernel scaffolding
Summary:
Adds support for raw CUDA kernels and compilation in wav2letter so we can open source FullConnectionCriterion/Viterbi CUDA kernels for ASG, etc. Adds CUB as a dependency, which is header only and automatically pulled in on build.

A few atrocities are present here to accomodate old versions of CMake:
- `cuda_add_library` is redefined to fix a bug with the version present in CMake 3.5.1 (and some slightly later versions): https://gitlab.kitware.com/cmake/cmake/issues/16602
- NVCC flags for C++ 11 must be explicitly set; host compilation flags can't be shared with CUDA compilation else very bad things happen
- `cub` headers are included in global nvcc scope. There isn't any other option
- `libw2lCriterionCudaKernels.a` now lives in `build/src/criterion` as output

Reviewed By: jcai1

Differential Revision: D13646966

fbshipit-source-id: a437350340152bec79dfe7d98e0c785ef8a6cf0d