Remove MKL has a hard dependency; rely on mklml if present
Summary:
As per https://github.com/facebookresearch/flashlight/issues/30, MKL-DNN can be linked to needed BLAS kernels (for gemm) in one of two ways: either directly with MKL (which gives better performance on many systems and int8/int16 gemm support in MKL-DNN with MKL >= 2018), or with a standalone library included with MKL-DNN called mklml.

MKL is technically not required for MKL-DNN, and it can be built without it/relying completely on mklml, which provides several implementations of necessary BLAS routines. Remove the dependency on MKL, and find and link MKLML if present.

Reviewed By: vineelpratap

Differential Revision: D13942646

fbshipit-source-id: a852c6847820084b737c2c4f33ffb7156bad62a9