Fix lexicon decoder for using '--surround' with CTC
Summary:
When using `--surround` in training, a silence token will be added in the very beginning of the transcription. This will cause trouble in CTC decoding for the following reason:
1. Decoding starts from lexicon root where it represents token silence.
2. CTC usually predicts blank tokens at the beginning.
3. One blank is predicted, the decoder will never be able to predict a silence token (the first token from --surround), because silence is not a child of lexicon root and decoder is not able to stay at the same lexicon node.

This diff is to solve point 3 by allowing the decoder to stay at the root. It also solves a small bug in `Decode.cpp`.

Reviewed By: jacobkahn, tlikhomanenko

Differential Revision: D23853720

fbshipit-source-id: e9b0b882d02773260770934bea3d6e031a077096